LayoutTests/imported/w3c/web-platform-tests/dom/events/event-global-set-before-handleEvent-lookup.any-expected.txt


PASS self.event is set before 'handleEvent' lookup
